Split utilities methods into multiple files
